refactor(DocThread): move salon files into respective component dirs
- Move article/brief/cover/outline_columns/outline_toc/stack/tile layout salons
  into their corresponding component directories
- Delete unused salon/lists_layout and salon/outline_layout (dead code)
