+++ frontend/core/unit/DocThread/ArticleLayout/NaviHead.tsx
@@ -1,8 +1,8 @@
 import ArrowButton from '~/widgets/Buttons/ArrowButton'
 
-import useSalon from '../salon/article_layout/navi_head'
 import useLogic from '../useLogic'
 import HeadAction from './HeadAction'
+import useSalon from './salon/navi_head'
 
 export default function NaviHead() {
   const s = useSalon()
