+++ frontend/core/unit/DocThread/ArticleLayout/salon/article_cover.ts
