+++ frontend/core/unit/DocThread/ArticleLayout/salon/head_action.ts
