+++ frontend/core/unit/DocThread/ArticleLayout/salon/toggle_btn.ts
