+++ frontend/core/unit/DocThread/BriefCardsLayout/Category.tsx
@@ -3,8 +3,8 @@ import type { FC } from 'react'
 import BookSVG from '~/icons/Book'
 import type { TArticle } from '~/spec'
 
-import useSalon from '../salon/brief_cards_layout/category'
 import useLogic from '../useLogic'
+import useSalon from './salon/category'
 
 type TProps = {
   title: string
