+++ frontend/core/unit/DocThread/BriefCardsLayout/salon/category.ts
