+++ frontend/core/unit/DocThread/BriefCardsLayout/salon/index.ts
