+++ frontend/core/unit/DocThread/CoverCardsLayout/salon/category.ts
