+++ frontend/core/unit/DocThread/OutlineColumnsLayout/index.tsx
@@ -1,7 +1,7 @@
 import { mockHelpCats } from '~/mock'
 
-import useSalon from '../salon/outline_columns_layout'
 import Category from './Category'
+import useSalon from './salon'
 
 export default function OutlineColumnsLayout() {
   const cats = mockHelpCats()
