+++ frontend/core/unit/DocThread/OutlineColumnsLayout/salon/category.ts
