+++ frontend/core/unit/DocThread/OutlineTocLayout/index.tsx
@@ -1,7 +1,7 @@
 import { mockHelpCats } from '~/mock'
 
-import useSalon from '../salon/outline_toc_layout'
 import Category from './Category'
+import useSalon from './salon'
 
 export default function OutlineTocLayout() {
   const cats = mockHelpCats()
