+++ frontend/core/unit/DocThread/StackCardsLayout/Category.tsx
@@ -5,8 +5,8 @@ import type { TArticle } from '~/spec'
 import ArrowButton from '~/widgets/Buttons/ArrowButton'
 import Facepile from '~/widgets/Facepile'
 
-import useSalon from '../salon/stack_cards_layout/category'
 import useLogic from '../useLogic'
+import useSalon from './salon/category'
 
 const FOLD_LIMIT = 5
