+++ frontend/core/unit/DocThread/StackCardsLayout/salon/category.ts
